wallet builds successfully
